applied feedback changes
