+++ src/types/hex-color-code.ts
@@ -2,9 +2,9 @@ import {StrongType, makeStrong} from '../strong-type';
 
 import {STRules} from '../rules';
 
-export type StrongHexCode = StrongType<string>;
+export type StrongHexColorCode = StrongType<string>;
 
-export function makeHexCode(initial: string | null | undefined, fallback: string): StrongHexCode {
+export function makeHexColorCode(initial: string | null | undefined, fallback: string): StrongHexColorCode {
 	const rules = new STRules();
 	rules.add().must.match.type.string();
 	rules.add().must.be.hexColorCode();
